Split the code into some smaller units
